% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metareg.R
\name{sim_metareg}
\alias{sim_metareg}
\title{sim_metareg}
\usage{
sim_metareg(
  min_obs = 100,
  max_obs = 500,
  n_studies = 100,
  intercept = 0.5,
  random = FALSE,
  tau = 0.01,
  formula,
  mod_types
)
}
\arguments{
\item{min_obs}{minimum number of observations per study}

\item{max_obs}{maximum number of observations per study}

\item{n_studies}{number of studies}

\item{intercept}{effect size (Hedges' g) without moderation effect}

\item{random}{choose TRUE to add variability of true effect size}

\item{tau}{standard deviation of true effect size}

\item{formula}{specify metaregression formula}

\item{mod_types}{types of predictor variables (categorical = 'cat', continuous = 'cont')}
}
\value{
data frame containing aggregated data
}
\description{
sim_metareg
}
\examples{
sim <- sim_metareg(random = TRUE,
tau = 0.05,
formula = y ~ -0.5*x1 + 1.2*x2,
mod_types = c('cat', 'cont'))
}
