% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_meta.R
\name{sim_meta}
\alias{sim_meta}
\title{sim_meta}
\usage{
sim_meta(
  min_obs,
  max_obs,
  n_studies,
  es_true,
  es = "SMD",
  fixed = TRUE,
  random = FALSE,
  tau = 0.01,
  varnames = c("x", "y"),
  n_variance = 0,
  metaregression = FALSE,
  mod_varname = c(),
  mod_labels = c(),
  mod_effect = NA
)
}
\arguments{
\item{min_obs}{minimum number of observations per study}

\item{max_obs}{maximum number of observations per study}

\item{n_studies}{number of studies}

\item{es_true}{true effect size}

\item{es}{type of effect size (choose 'SMD' fpr standardized mean difference and 'ZCOR' for correlations)}

\item{fixed}{simulation based on fixed effects model (TRUE)}

\item{random}{simulation based on between study heterogenity (random-effects model)}

\item{tau}{specify standard deviation of between study heterogenity. If moderator variable specified, tau is the residual heterogenity after including the moderator.}

\item{varnames}{variable names}

\item{n_variance}{variance between n1 and n2. Both sample sizes are drawn from normal distributions using the specified value as standard deviation.}

\item{metaregression}{set TRUE if categorical or continuous moderator should be included.}

\item{mod_varname}{specify the variable name of the moderator}

\item{mod_labels}{specify the labels of different groups of studies, if the moderator is a categorical variable.}

\item{mod_effect}{standardize size of moderation effect for categorical moderator. Provide a vector of multiple effect sizes if moderator variable includes more than two groups.}
}
\value{
list containing raw data (data_raw) and aggregated data with computed effects sizes and standard errors (data_aggr)
}
\description{
sim_meta
}
\examples{
sim_meta(min_obs = 100, max_obs = 500, n_studies = 20, es_true = 0.7)
}
