% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_meta.R
\name{sim_meta}
\alias{sim_meta}
\title{sim_meta}
\usage{
sim_meta(
  min_obs = 100,
  max_obs = 500,
  n_studies = 100,
  smd_true = 0.5,
  r_true = 0,
  sd1 = 1,
  sd2 = 1,
  random = FALSE,
  tau = 0.01,
  metaregression = FALSE,
  mod_labels = c(),
  smd_mod_effects = c(),
  r_mod_effects = c(),
  aggregate = TRUE,
  random_effects = c()
)
}
\arguments{
\item{min_obs}{minimum number of observations per study}

\item{max_obs}{maximum number of observations per study}

\item{n_studies}{number of studies}

\item{smd_true}{true effect size of standardized mean difference between variable x and y}

\item{r_true}{correlation between x and y. Has to be specified if effect size is with-groups (correlation/ pre-post mean difference)}

\item{sd1}{standard deviation of x}

\item{sd2}{standard deviation of y}

\item{random}{if \code{TRUE} simulation includes between study heterogenity (random-effects model), if \code{FALSE} fixed effects model is used}

\item{tau}{specify standard deviation of between study heterogenity. If moderator variable specified, tau is the residual heterogenity after including the moderator.}

\item{metaregression}{set TRUE if categorical moderator should be included.}

\item{mod_labels}{labels of different subgroups of studies, if the moderator is a categorical variable.}

\item{smd_mod_effects}{effect sizes in subgroups. Provide a vector of multiple effect sizes if moderator variable includes more than one group.}

\item{r_mod_effects}{effect on correlation in each subgroup.}

\item{aggregate}{return only aggregated data (TRUE) or raw data (FALSE).}

\item{random_effects}{Choose if 'SMD' or 'ZCOR' should be include random effects}
}
\value{
list containing raw data (data_raw) and aggregated data with computed effects sizes and standard errors (data_aggr)
}
\description{
sim_meta
}
\examples{
sim_meta(min_obs = 100, max_obs = 500, n_studies = 20, smd_true = 0.7)
}
