% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_meta.R
\name{sim_meta}
\alias{sim_meta}
\title{sim_meta}
\usage{
sim_meta(
  min_obs,
  max_obs,
  n_studies,
  es_true,
  es = "SMD",
  fixed = TRUE,
  random = FALSE,
  tau = 0.01,
  varnames = c("x", "y")
)
}
\arguments{
\item{min_obs}{minimum number of observations per study}

\item{max_obs}{maximum number of observations per study}

\item{n_studies}{number of studies}

\item{es_true}{true effect size}

\item{es}{type of effect size (choose 'SMD' fpr standardized mean difference and 'ZCOR' for correlations)}

\item{fixed}{simulation based on fixed effects model (TRUE)}

\item{random}{simulation based on between study heterogenity (random-effects model)}

\item{tau}{specify standard deviation of between study heterogenity}

\item{varnames}{variable names}
}
\value{
list containing raw data (data_raw) and aggregated data with computed effects sizes and standard errors (data_aggr)
}
\description{
sim_meta
}
\examples{
sim_meta(min_obs = 100, max_obs = 500, n_studies = 20, es_true = 0.7)

}
